Add an HTML/JavaScript drop down menu option to parserPopUp.pl.
The point of this is to make a drop down menu that can contain math mode
content. A native select of course can not contain such content other
than in string form which is ugly at best.

The option is `useHTMLSelect`.  The default value for this option is 1
which means that a native HTML select element is used.  That means that
the current behavior of a drop down menu is used. If `useHTMLSelect` is
set to 0, then instead a Bootstrap drop down is used instead of a native
HTML select. In this case the choices must be provided that satisfy the
constraint that they will work directly in HTML and will also work if
placed in a `\text{...}` call in LaTeX.  In HTML of course `\(...\)` or
`\[...\]` will work and will be typeset by MathJax.  Those also work
inside `\text{...}` in LaXeX.  So the drop down could have choices like
`\(y < \frac{3}{4}\)` or `Choice 1: \(y^2 = e^x\)`.

The drop down menu is styled to appear much like the native select, and
JavaScript designed to make the drop down behave much the same. There
are some differences such as the down arrow for a Bootstrap drop down
menu looks a little different than that of a select element, and there
is a hover/focus background color change for a Bootstrap drop down.