Fix the values option of parserPopUp.pl.
The current code sets the correct choice to be the label, and it should
be the value.  As such if the values are specified, none of the answers
are accepted as correct because the correct label is not equal to the
correct value.

Here is a minimal test case:

```perl
DOCUMENT();
loadMacros(qw(PGstandard.pl PGML.pl parserPopUp.pl PGcourse.pl));

$popup = DropDown([ 'first', 'second', 'third' ], 1, values => [ 'a', 'b', 'c' ]);

BEGIN_PGML
Select "second": [_]{$popup}
END_PGML

ENDDOCUMENT();
```

On the develop branch (or main) no answers will be accepted with this
problem.  With this pull request the second answer will be accepted as
correct.  If you delete the `values` option from the `DropDown` call,
then the second answer will be accepted as correct with the develop
branch (and main), and with this pull request.