Fix a typo in the `data_feedback_insert_element` attribute and use a div instead of a span.
This is not particularly consequential since the feedback button is
still inserted after the select.  That is the same as what the data
attributes end up doing.

However, using a `span` is a problem, becuase the feedback popover is
inserted into the span which results in invalid HTML. So a display
inline `div` is used instead.