Fix a couple of issues seen in review.
Return true from the sineWaveTool updateHighlights method so that the
general graphtool move handler doesn't overrided the coordinates set in
the updateHighlights method.

During the sine wave tool graphing phases instead of refusing to plot a
point when a down event occurs and the cursor is in an invalid location,
just plot the point where the highlight point currently is. Refusing to
plot the point can be confusing because the highlight point looks like
it is where the user wants it to be, and yet the click does nothing.
This technique needs to be implemented for all of the tools. It is
better than the previous method since it is more intuitive for the user.