Make changes for existing graph tools that were made with the sine wave tool.
For all of the graph tools switch to using the highlight point
coordinates when a click occurs at an invalid location on the board for
the next phase.  Still refuse to create a point if a click or touch
occurs off the board.  This doesn't really happen for a click, but does
for a touch on a touch screen device.  In that case, the touch off the
board should be ignored.

Don't use `==` for comparison of floating point numbers.