Add a scaffold `preview_can_change_state` option that determines behavior when preview is used.
The default value is 1, and in that case scaffolds will continue to
behave as they currently do.  If this option is set to 0, then when an
answer preview occurs, the scaffold state will remain the same as before
the preview occured.  Opened scaffolds will stay open, closed scaffolds
will stay closed, and scaffolds that could't be opened still can't be
opened.  Note this refers to the initial open/closed state when the
problem loads, and does not respect opening or closing of scaffolds
(that can be opened or closed) by the user.

This uses the persistence hash to store the state. Note that state is
really the scores for all answers in scaffold sections.  It is important
to note that state is usually not saved by the frontend until an answer
submission occurs (although it is saved in a hidden form field for
instructors), and if there is no state, then it assumes the problem is
in its initial state, and uses scores of 0 for all answers.