Give explicit multiplication the special mq_precedence as well.
Implied multiplication ` *` currently has the precedence of 2.9 when
MathQuill is used.  This means that `(x-3)/4 (x+8)/7` is interpreted as
`\frac{(x-3)}{4} * \frac{(x+8)}{7}`.

However, `(x-3)/4 * (x+8)/7` is interpreted as
`\frac{\frac{(x-3)/4 * (x+8)}{7}`.  That should also be interpreted as
`\frac{(x-3)}{4} * \frac{(x+8)}{7}` and is with the mq_precendence for
`* ` also set to 2.9.