Add the listAnswer.pl macro.
This macro was originally omitted because it is not directly used by any
problems in the OPL or Contrib.  However, the CofIdaho_macros.pl macro
loads it if the `Picky_equation_cmp` method is used.  So any problem
that used that macro and method don't work right.  Apparently several of
the problems in the `Library/CollegeOfIdaho/setAlgebra_01_05_ProblemSolvingAndFormulas`
use this.

See https://webwork.maa.org/moodle/mod/forum/discuss.php?d=8632#p21807.