Fix coordinate hints during any restricted point drag.
Any time that a point's drag is restricted the coordinate hints can be
wrong. This happens if the drag is in a range that the point can't be
moved to and can happen for both mouse or keyboard drags.

For an extreme example where this is easy to see, with the 4-point cubic
tool graph the points (-1, 0), (0, 0), (1, 0), and (2, 0).  Then click
on the point at (2, 0) and slowly drag it to the left. Without this pull
request you will see the coordinate hints show (1, 0), then (0, 0),
etc., even though that is not where the point being dragged is.