If `$showPartialCorrectAnswers = 0` and all answers are correct, then show all correct.
This makes it so that if `$showPartialCorrectAnswers = 0` and all
answers are correct, then the feedback buttons are all green revealing
that all answers are correct.

This works by passing the problem state found when evaluating answers
into the post content processor methods (it is a new third parameter
passed to the callback). The ENDDOCUMENT post processor then uses this
to quickly determine that the score for the problem is 1 (or more), and
in that case shows the usual feedback showing that all answers are
correct.

This avoids looping multiple times to determine correctness of the
problem.