Add a vector tool to the graph tool.
This was requested by @somiaj in our last developer meeting after I told
him that I already had this encoded. I have had it as a custom tool for
a long time now.

The usage of the new tool is documented in the parserGraphTool.pl
macro.

The "vector" graph object derives from the "line" graph object, and the
"VectorTool" tool derives from the "LineTool" tool.  That saves some
code redundancy.

Note that there will be some minor conflicts with #1191 that will need
to be resolved.