Make crossed quadrilaterals fill only the region containing the fill point.
This is to fix a comment I placed in the code about filling
quadrilaterals, and the comment made by @Alex-Jordan in
https://github.com/openwebwork/pg/pull/1191#issuecomment-2683381959.

The code implemented in #1191 fills the entire interior of a
quadrilateral even if it is a crossed quadrilateral (that is a
quadrilateral with two edges that intersect). That is inconsistent with
all of the other graph tool objects in some sense.  To be more
consistent only the region inside the quadrilateral that satisfies the
same inequalities relative to the borders should be filled.  So that is
what this does.

I also was doing some benchmarking of the TikZ fill code used for
generating hardcopy and noticed that some things were rather slow. The
slow down it turns out was using MathObject Reals in computations. There
is no need for the MathObject structure in those computations, so all of
the fill code now uses the perl values instead. It isn't even funny how
much faster computations are with pure perl numbers versus MathObject
Reals. The difference is rather extreme. Performing something like 10000
basic arithmetic operations with MathObject Reals takes more than five
seconds, but with pure perl numbers it takes less than 20 milliseconds.