Add a segment tool and graph object and non positional vector grading option.
The segment tool and vector tool both derive from the line tool and
share much of the same code. The segment graph object and vector graph
object derive from the line graph object and also share much of the same
code.

Also add the `vectorsArePositional` option.  This is 0 by default which
means that if a vector is graphed that has the same magnitude and
direction as the correct answer, then it is considered correct. If the
option is set to one, then a vector must also have the same initial
point as the correct answer to be considered correct.

Note that if the `vectorsArePositional` option is 0 and a student graphs
the same vector twice at different locations, then the second vector
will not be considered correct. That is unless there are two correct
answers that are also the same vector.

However, if the student graphs the same vector twice at the same
location (so visibly only one is seen), then the second vector is
ignored, and the answer will still be correct.  This follows the usual
rule of the default grader in attempting to grade things according to
what is seen.