Switch to a better extension system for the GraphTool.
For JavaScript instead of piecing together the parts of the class
definition, use a function that returns the class definition directly.
This has the advantage of being able to use `super` directly to access
methods of the parent class, and is just much easier to work with.

All of the graph objects and graph tools are now in separate files
(except the SelectTool and SolidDashTool which are core functionality).

For Perl, all graph objects are packages that derive from the
GraphTool::GraphObject package instead of the previous method of
defining cmp and tikz subroutines in a hash. A mapping from the string
identifier to the package is maintained and used to construct the
correct perl object for a graph tool object. There is a compatibility
layer for objects added the old way.