Fix filling of concave quadrilaterals.
Currently only the portion of a concave quadrilateral that the fill
point is in is filled, and not the entire interior because the current
check separates portions that satisfy different inequalities with
respect to the borders.  So this makes it so the interior parts are only
distinquished for crossed quadrilaterals.

Also fix the documentation about how to use the `%graphToolObjectCmps`
hash entry for vectors.  The GraphTool object must be passed so that it
can get the `vectorsArePositional` option. This is only for those using
this for a custom checker (if anyone does).

Also fix an incorrect `this` usage for intervals. This makes it possible
to drag endpoints of an interval onto each other.

This was included in #1206.  However, I am (at least temporarily)
changing that pull request to a draft until (and if) some issues are
resolved.  These things are unrelated to that pull request really.