Make a single `onBoundary` method to each object.
This method is passed what is needed and returns if the given point is
on the boundary of the object or not.

Vectors and segments check to see if they are crossed when moving from
one check point to another when trying to determine if a fill check
point is a border pixel. This is finally a reliable approach for
segments and vectors.