Fix an infinite loop that is occurring unexplicably.
This is a quick hack because I can't figure this out.  When certain
segments or vectors are graphed that poke into the region or don't bound
the region, then the flood fill algorithm starts repeatedly filling
already filled pixels in an infinite loop. The break added here stops
the loop and the fill completes seemingly correctly.  The corresponding
Perl algorithm does not have this problem.