Switch the default back to the inequality fill for the graph tool.
Make the flood fill algorithm opt in.  This is probably better and is
certain to be backward compatibe.

Perhaps auto detection could be used and the flood fill algorithm
switched to if certain objects are graphed that the flood fill algorithm
works better for.  However, I am going with this for now.