Fix a vulnerability in `Statistics.pm` and the macro `PGstatisticsmacros.pl`.
This vulnerability allows a problem to write to any location or
overwrite any file that the server has write permission to.

The problem is that the `write_array_to_CSV` method of the
`Statistics.pm` module takes a file name as its first argument and
writes to that file without question.

To fix this the method instead uses the new `WeBWorK::PG::IO::saveDataToFile`
method which refuses to write anything outside of the html temporary
directory.

Note that the `make_csv_alias` method of the `Statistics.pm` module was
dropped because it uses an invalid approach to creating a unique file
name and is not needed.  So there is no need for the `Statistics`
package to even be an object.  Now there is just the method
`write_array_to_CSV`.

Previously the `insertDataLink` method of `PGstatisticsmacros.pl`
accepted the `$PG` object for its first argument.  That first argument
is still accepted, but is dropped if given, and there is no need to pass
that argument anymore. The macro has direct access to the `$PG` variable
so it is not a needed parameter.

Also, a new optional last argument to the `insertDataLink` is accepted.
If provide this argument must be a reference to a hash, and the
key/value pairs in that hash will be added as attributes to the HTML `a`
tak that is returned. If this parameter is not provided or it is but
does not contain a `download` attribute, then a default `download`
attribute with value `data.csv` will be added.

Note that the `insertDataLink` method is not used by any problem in the
OPL or Contrib.  Furthermore, nothing in the `Statistics.pm` module is
directly used by any problem in OPL or Contrib. Backwards compatibility
is maintained for the `insertDataLink` method for anyone that may have
been using it, but not for the `Statistics.pm` module.  It was not
intended that the module be used directly anyway.

This fixes the second of the three ways that I know of for a problem to
directly write to anywhere that the server has permission to do so.