Fix comparison of a fraction to infinity.
If an infinity is the left operand and a `Fraction` the right operand of
an inequality operator, then the operator is returning the wrong thing.
This is caused by the change in #1208 failing to account for the op
order flag when a fraction is compared to an infinity.

For example if `$a = Compute('5 / 3')` (in the `Fraction` context),
`$posInf = Compute('-inf')`, and `$negInf = Compute('inf')`, then

`$posInf < $a` is true,
`$posInf > $a` is false,
`$negInf < $a` is false, and
`$negInf > $a` is true.

All of which are exactly the opposite of the correct result.

However,

`$a < $posInf` is true,
`$a > $posInf` is false,
`$a < $negInf` is false, and
`$a > $negInf` is true

as is expected.

The problem is that since the op order flag is not accounted for, the
first four comparisons are really returning the result of the latter
four comparisons.

With this pull request all of the above comparisons are correct.